# PS back/sprint/openapi.yaml
openapi: 3.0.0
info:
  title:  Gestão de Montagem de Drones
  version: 1.0.0
paths:
...
  /utilizador/montagem:
    post:
      tags:
        - Montagem
      summary: Adicionar uma nova montagem
      description: Adiciona uma nova montagem à lista de montagens 
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPut'
      responses:
        '201':
          description: Montagem adiconada com sucesso!
        '400':
          description: Requisição inválida
  /utilizador/{idm}/{idp}:
    get:
      tags:
        - Peças
      summary: Obter peça através do ID
      description: Obtem dados deuma peça de uma montagem
      parameters:
        - name: idm
          in: path
          description: ID da montagem
          required: true
          schema:
            type: integer
            format: int64
        - name: idp
          in: path
          description: ID da peça
          required: true
          schema:
            type: integer
            format: int64    
      responses:
        '200':
          description: Peça encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PecaGet'
        '404':
          description: Montagem ou peça não existem
    delete:
      tags:
        - Peças
      summary: Remover peça através do ID
      description: Remove uma peça conforme ID
      parameters:
        - name: idm
          in: path
          description: ID da montagem
          required: true
          schema:
            type: integer
            format: int64
        - name: idp
          in: path
          description: ID da peça
          required: true
          schema:
            type: integer
            format: int64    
      responses:
        '204':
          description: Peça desmontada!
        '404':
          description: Montagem ou peça não existem
    put:
      tags:
        - Peças
      summary: Atualizar peça através do ID
      description: Atualiza dados especificos de uma peça
      parameters:
        - name: idm
          in: path
          description: ID da montagem
          required: true
          schema:
            type: integer
            format: int64
        - name: idp
          in: path
          description: ID da peça
          required: true
          schema:
            type: integer
            format: int64    
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PecaGet'
      responses:
        '200':
          description: Peça atualizada com sucesso
        '400':
          description: Requisição inválida
        '404':
          description: Montagem ou peça não existem
...
components:
  schemas:
...
    UserPut:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        descrption:
          type: array
          items:
            $ref: '#/components/schemas/Item'
      required:
        - id
        - name
        - descrption
      example:
        id: 123
        name: Drone favorito
        description: Modelo X, feito por Y
    PecaGet:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: array
          items:
            $ref: '#/components/schemas/Item'
        peca:
          type: string
      required:
        - idm
        - idp
      example:
        id: 456
        name: Drone X
        items:
          - peca: hélice 12/06/2023
          - peca: camara 31/05/2023
          - peca: corpo 4/02/2023
          - peca: motor 15/06/2022
...
    Item:
      type: object
      properties:
        peso:
          type: integer
        nome:
          type: string